requirements.txt
cython==0.29.14
scipy==1.4.1
numpy==1.18.1
tensorflow==2.1.0
keras==2.3.1
utils==1.0.0
opencv-python==4.1.2.30
opencv-contrib-python==4.2.0.32
imgaug==0.3.0
pyyaml==5.1.2

